ensure run with sudo

+++ update-docker-pihole.sh
@@ -1,4 +1,9 @@
 #!/bash
+if [[ $(id -u) -ne 0 ]]; then
+  echo "usage: sudo sh update-docker-pihole.sh"
+  exit 1
+fi
+
 echo -e "nameserver\t1.1.1.1" | sudo tee /etc/resolv.conf
 docker pull pihole/pihole
 docker rm -f pihole
